requirements.txt
streamlit 
requests 
beautifulsoup4 
pandas 
sentence-transformers 
faiss-cpu 
pillow